# frontend/web/src/api/openapi.json
{
 "components": {
  "schemas": {
...
   "AnalyticsEventType": {
    "enum": [
     "view",
     "read_complete"
    ],
    "title": "AnalyticsEventType",
    "type": "string"
   },
...
   "Beacon": {
    "properties": {
     "article_id": {
      "format": "uuid",
      "title": "Article Id",
      "type": "string"
     },
     "event_type": {
      "$ref": "#/components/schemas/AnalyticsEventType"
     },
     "lang": {
      "maxLength": 10,
      "pattern": "^[a-z]{2}(-[A-Z][A-Za-z]{1,3})?$",
      "title": "Lang",
      "type": "string"
     },
     "session_hash": {
      "description": "A random id the reader's browser makes each day; nothing about the reader.",
      "pattern": "^[0-9a-f]{16,64}$",
      "title": "Session Hash",
      "type": "string"
     }
    },
    "required": [
     "article_id",
     "lang",
     "event_type",
     "session_hash"
    ],
    "title": "Beacon",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "PublicArticle": {
    "properties": {
     "access": {
      "default": "free",
      "title": "Access",
      "type": "string"
     },
     "article_id": {
      "format": "uuid",
      "title": "Article Id",
      "type": "string"
     },
     "blocks": {
      "items": {
       "$ref": "#/components/schemas/PublicBlock"
      },
      "title": "Blocks",
      "type": "array"
     },
     "company": {
      "title": "Company",
      "type": "string"
     },
     "company_id": {
      "format": "uuid",
      "title": "Company Id",
      "type": "string"
     },
     "lang": {
      "title": "Lang",
      "type": "string"
     },
     "langs": {
      "additionalProperties": {
       "type": "string"
      },
      "title": "Langs",
      "type": "object"
     },
     "locked": {
      "default": false,
      "title": "Locked",
      "type": "boolean"
     },
     "path": {
      "title": "Path",
      "type": "string"
     },
     "published_at": {
      "format": "date-time",
      "title": "Published At",
      "type": "string"
     },
     "slug": {
      "title": "Slug",
      "type": "string"
     },
     "sources": {
      "items": {
       "$ref": "#/components/schemas/PublicSource"
      },
      "title": "Sources",
      "type": "array"
     },
     "summary": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Summary"
     },
     "title": {
      "title": "Title",
      "type": "string"
     }
    },
    "required": [
     "article_id",
     "lang",
     "slug",
     "path",
     "title",
     "summary",
     "published_at",
     "blocks",
     "sources",
     "langs",
     "company",
     "company_id"
    ],
    "title": "PublicArticle",
    "type": "object"
   },
   "PublicArticleSummary": {
    "properties": {
     "access": {
      "default": "free",
      "title": "Access",
      "type": "string"
     },
     "article_id": {
      "format": "uuid",
      "title": "Article Id",
      "type": "string"
     },
     "lang": {
      "title": "Lang",
      "type": "string"
     },
     "path": {
      "title": "Path",
      "type": "string"
     },
     "published_at": {
      "format": "date-time",
      "title": "Published At",
      "type": "string"
     },
     "slug": {
      "title": "Slug",
      "type": "string"
     },
     "summary": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Summary"
     },
     "title": {
      "title": "Title",
      "type": "string"
     }
    },
    "required": [
     "article_id",
     "lang",
     "slug",
     "path",
     "title",
     "summary",
     "published_at"
    ],
    "title": "PublicArticleSummary",
    "type": "object"
   },
   "PublicBlock": {
    "properties": {
     "text": {
      "title": "Text",
      "type": "string"
     },
     "type": {
      "title": "Type",
      "type": "string"
     }
    },
    "required": [
     "type",
     "text"
    ],
    "title": "PublicBlock",
    "type": "object"
   },
   "PublicSource": {
    "properties": {
     "site": {
      "title": "Site",
      "type": "string"
     },
     "title": {
      "title": "Title",
      "type": "string"
     },
     "url": {
      "title": "Url",
      "type": "string"
     }
    },
    "required": [
     "title",
     "site",
     "url"
    ],
    "title": "PublicSource",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "scheme": "bearer",
    "type": "http"
   }
  }
 },
 "info": {
  "title": "Autora API",
  "version": "0.0.1"
 },
 "openapi": "3.1.0",
 "paths": {
  "/api/analytics/beacon": {
   "post": {
    "description": "Count a view or a completed read. A repeat from the same session that day is dropped\n(still 204: the reader's page has nothing to do about it).",
    "operationId": "beacon_api_analytics_beacon_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/Beacon"
       }
      }
     },
     "required": true
    },
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Beacon",
    "tags": [
     "public"
    ]
   }
  },
...
  "/api/public/articles": {
   "get": {
    "operationId": "list_articles_api_public_articles_get",
    "parameters": [
     {
      "in": "query",
      "name": "lang",
      "required": true,
      "schema": {
       "maxLength": 10,
       "pattern": "^[a-z]{2}(-[A-Z][A-Za-z]{1,3})?$",
       "title": "Lang",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "company",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "maxLength": 100,
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Company"
      }
     },
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 20,
       "maximum": 50,
       "minimum": 1,
       "title": "Limit",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/PublicArticleSummary"
         },
         "title": "Response List Articles Api Public Articles Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Articles",
    "tags": [
     "public"
    ]
   }
  },
  "/api/public/articles/{lang}/{slug}": {
   "get": {
    "operationId": "get_article_api_public_articles__lang___slug__get",
    "parameters": [
     {
      "in": "path",
      "name": "lang",
      "required": true,
      "schema": {
       "title": "Lang",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "slug",
      "required": true,
      "schema": {
       "title": "Slug",
       "type": "string"
      }
     },
     {
      "in": "cookie",
      "name": "autora_reader",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Autora Reader"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PublicArticle"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Article",
    "tags": [
     "public"
    ]
   }
  },
...
 }
}